Fix out of array bounds

+++ engine/common/filesystem.c
@@ -3264,20 +3264,24 @@ static dlumpinfo_t *W_FindLump( wfile_t *wad, const char *name, const char match
 
 	// trying to extract hint from the name
 	FS_FileBase( name, barename );
-	Q_strncpy( suffix, barename + Q_strlen( barename ) - HINT_NAMELEN, sizeof( suffix ));
 
-	// we not known about filetype, so match only by filename
-	for( hint = wad_hints; hint->ext; hint++ )
+	if( Q_strlen( barename ) >= HINT_NAMELEN )
 	{
-		if( !Q_stricmp( suffix, hint->ext ))
-		{
-			img_type = hint->type;
-			break;
-		}
-	}
+		Q_strncpy( suffix, barename + Q_strlen( barename ) - HINT_NAMELEN, sizeof( suffix ));
 
-	if( img_type != IMG_DIFFUSE )
-		barename[Q_strlen( barename ) - HINT_NAMELEN] = '\0'; // kill the suffix
+		// we not known about filetype, so match only by filename
+		for( hint = wad_hints; hint->ext; hint++ )
+		{
+			if( !Q_stricmp( suffix, hint->ext ))
+			{
+				img_type = hint->type;
+				break;
+			}
+		}
+
+		if( img_type != IMG_DIFFUSE )
+			barename[Q_strlen( barename ) - HINT_NAMELEN] = '\0'; // kill the suffix
+	}
 
 	// look for the file (binary search)
 	left = 0;
